\name{proptimr}
\alias{proptimr}
\title{Compact display of an \code{optimr()} result object}
\description{
   \code{proptimr} displays the contents of a result computed by \code{optimr()}. 
}
\usage{
   proptimr(opres, nlim=5)
}
\arguments{
    \item{opres}{
        the object returned by function \code{optimr()}
        }
    \item{nlim}{
        Only the first nlim elements of the solution are displayed.
        }
}
\value{
This function is intended for output only.
}
\author{
    John C. Nash
}
\keyword{optimize}

