\name{optimx-package}
\alias{optimx-package}
\docType{package}
\title{
A replacement and extension of the optim() function
}
\description{
Provides a replacement and extension of the optim() function to unify
and streamline optimization capabilities in R for smooth, possibly box
constrained functions of several or many parameters
}
\details{
\tabular{ll}{
Package: \tab optimx\cr
Version: \tab 2011-7.10\cr
Date: \tab 2011-07-10\cr
License: \tab GPL-2\cr
Lazyload: \tab Yes\cr
Depends: \tab numDeriv, BB, ucminf, Rcgmin, Rvmmin, minqa, setRNG, kktc, useroptfn, bmchk\cr
Repository: \tab R-Forge\cr
Repository/R-Forge/Project: \tab optimizer\cr
Repository/R-Forge/Revision: \tab 478\cr
Date/Publication: \tab 2011-07-21 02:13:01\cr
Packaged: \tab 2011-07-26 21:17:43 UTC; rforge\cr
Built: \tab R 2.13.1; ; 2011-07-28 14:38:19 UTC; unix\cr
}

Index:
\preformatted{
get.result              Result extraction tool for answers from optimx
optansout               Output tool for answers from optimx
optimx                  General-purpose optimization
}
}
\author{
John C Nash <nashjc@uottawa.ca> and Ravi Varadhan <RVaradhan@jhmi.edu>

Maintainer: John C Nash <nashjc@uottawa.ca>
}
\references{
  Nash, John C. and Varadhan, Ravi (2011) Unifying Optimization Algorithms 
          to Aid Software System Users: optimx for R, Journal of Statistical
	  Software, publication pending.
}
\keyword{ package }
\keyword{ optimization }
\seealso{
optimx
}

