\name{optimx-package}
\alias{optimx-package}
\docType{package}
\title{
A replacement and extension of the optim() function
}
\description{
Provides a replacement and extension of the optim() function to unify
and streamline optimization capabilities in R for smooth, possibly box
constrained functions of several or many parameters
}
\details{
\tabular{ll}{
Package: \tab optimx\cr
Version: \tab 2012-2.202\cr
Date: \tab 2012-03-16\cr
License: \tab GPL-2\cr
Lazyload: \tab Yes\cr
Depends: \tab numDeriv, optfntools\cr
Suggests: \tab BB, ucminf, Rcgmin, Rvmmin, minqa, setRNG, dfoptim\cr
Repository: \tab R-Forge\cr
Repository/R-Forge/Project: \tab optimizer\cr
}

Index:
\preformatted{
get.result              Result extraction tool for answers from optimx
optansout               Output tool for answers from optimx
optimx                  General-purpose optimization
}
}
\author{
John C Nash <nashjc@uottawa.ca> and Ravi Varadhan <RVaradhan@jhmi.edu>

Maintainer: John C Nash <nashjc@uottawa.ca>
}
\references{
  Nash, John C. and Varadhan, Ravi (2011) Unifying Optimization Algorithms 
          to Aid Software System Users: optimx for R, Journal of Statistical
	  Software, publication pending.
}
\keyword{ package }
\keyword{ optimization }
\seealso{
optimx
}

